refactor: page to ssr; use DS styling
