+++ app/[locale]/collectibles/_components/CollectiblesContent/loading.tsx
@@ -1,0 +1,5 @@
+// TODO: Loading placeholder
+
+const Loading = () => <div />
+
+export default Loading
