+++ app/[locale]/collectibles/_components/collectibles.tsx
@@ -1,34 +1,30 @@
 "use client"
 
 import React from "react"
+import { useLocale } from "next-intl"
 import { WagmiProvider } from "wagmi"
 import { type Locale, RainbowKitProvider } from "@rainbow-me/rainbowkit"
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query"
 
-import { Badge, Stats } from "../page"
+import { Badge } from "../types"
 
-import CollectiblesContent from "./CollectiblesContent"
+import CollectiblesContent from "./CollectiblesContent/lazy"
 
 import { rainbowkitConfig } from "@/config/rainbow-kit"
 
-export interface CollectiblesPageProps {
+export type CollectiblesPageProps = {
   badges: Badge[]
-  stats: Stats
-  locale: string
 }
 
-const CollectiblesPage: React.FC<CollectiblesPageProps> = ({
-  badges,
-  stats,
-  locale,
-}) => {
+const CollectiblesPage = ({ badges }: CollectiblesPageProps) => {
   const queryClient = new QueryClient()
+  const locale = useLocale()
 
   return (
     <WagmiProvider config={rainbowkitConfig}>
       <QueryClientProvider client={queryClient}>
         <RainbowKitProvider locale={locale as Locale}>
-          <CollectiblesContent badges={badges} stats={stats} locale={locale} />
+          <CollectiblesContent badges={badges} />
         </RainbowKitProvider>
       </QueryClientProvider>
     </WagmiProvider>
