+++ app/[locale]/collectibles/constants.ts
@@ -1,0 +1,2 @@
+export const COLLECTIBLES_BASE_URL =
+  "https://ethereum-org-collectibles.vercel.app"
